+++ app/src/main/java/com/mohannad/coupon/view/adapter/product/ProductAdapter.java
@@ -114,7 +114,7 @@ public class ProductAdapter extends RecyclerView.Adapter<ProductAdapter.ProductV
             // desc
             this.itemView.tvDescItemProductRv.setText(product.getDesc());
             this.itemView.tvNumReviewsItemProductRv.setText(Integer.toString(product.getNumberReviews()));
-            this.itemView.tvReviewsItemProductRv.setText(Integer.toString(product.getRate()));
+            this.itemView.tvReviewsItemProductRv.setText(Float.toString(product.getRate()));
             this.itemView.tvNumOrderItemProductRv.setText(Integer.toString(product.getNumberOrders()));
 
             // check if the product in favorite or not
